requirements.txt
python-telegram-bot==20.3
torch==2.0.1
torchvision==0.15.2
numpy==1.21.6
opencv-python==4.8.0
Pillow==9.5.0
mediapipe==0.9.4